move encryption inside SignedEncryptedEvent
