+++ interwallet/src/shared/EncryptedEvent.ts
@@ -1,8 +1,13 @@
-import { bytesToHex, getAddress } from "viem"
+import { bytesToHex, getAddress, hexToBytes } from "viem"
+
+import { decode, encode } from "microcbor"
+import { base58btc } from "multiformats/bases/base58"
+import nacl from "tweetnacl"
 
 import * as Messages from "./messages.js"
-import { Room } from "./types"
-import { assert } from "."
+import { PrivateUserRegistration, Room, RoomEvent } from "./types"
+import { assert } from "./utils.js"
+import { equals } from "uint8arrays"
 
 type EncryptedPayload = {
 	publicKey: Uint8Array
@@ -62,5 +67,64 @@ export class EncryptedEvent {
 			senderPublicKey: event.senderPublicKey,
 			recipients: event.recipients,
 		})
+	}
+
+	static encryptMessageForRoom(room: Room, message: string, user: PrivateUserRegistration): EncryptedEvent {
+		const event = {
+			type: "message",
+			detail: { content: message, sender: user.address, timestamp: Date.now() },
+		}
+
+		const otherRoomMembers = room.members.filter(({ address }) => user.address !== address)
+		assert(otherRoomMembers.length > 0, "room has no other members")
+
+		const encodedEvent = encode(event)
+
+		return new EncryptedEvent(
+			base58btc.baseDecode(room.id),
+			hexToBytes(user.address),
+			hexToBytes(user.keyBundle.encryptionPublicKey),
+			otherRoomMembers.map((otherRoomMember) => {
+				const publicKey = hexToBytes(otherRoomMember.keyBundle.encryptionPublicKey)
+				const nonce = nacl.randomBytes(nacl.box.nonceLength)
+				const ciphertext = nacl.box(encodedEvent, nonce, publicKey, hexToBytes(user.encryptionPrivateKey))
+
+				return {
+					publicKey,
+					ciphertext,
+					nonce,
+				}
+			})
+		)
+	}
+
+	static decrypt(encryptedEvent: EncryptedEvent, user: PrivateUserRegistration): RoomEvent {
+		let messageToDecrypt: Messages.EncryptedPayload
+		let publicKey: Uint8Array
+
+		if (equals(encryptedEvent.senderAddress, hexToBytes(user.address))) {
+			// this user is the sender
+			// decrypt an arbitrary message, so choose the first one
+			messageToDecrypt = encryptedEvent.recipients[0]
+			publicKey = messageToDecrypt.publicKey
+		} else {
+			// otherwise this user is one of the recipients
+			const retrievedMessageToDecrypt = encryptedEvent.recipients.find(({ publicKey }) =>
+				equals(publicKey, hexToBytes(user.keyBundle.encryptionPublicKey))
+			)
+			assert(retrievedMessageToDecrypt !== undefined, "failed to find encrypted message for this user")
+			messageToDecrypt = retrievedMessageToDecrypt
+			publicKey = encryptedEvent.senderPublicKey
+		}
+
+		const decryptedEvent = nacl.box.open(
+			messageToDecrypt.ciphertext,
+			messageToDecrypt.nonce,
+			publicKey,
+			hexToBytes(user.encryptionPrivateKey)
+		)
+		assert(decryptedEvent !== null, "failed to decrypt room event")
+
+		return decode(decryptedEvent) as RoomEvent
 	}
 }
