+++ interwallet/src/shared/cryptography.ts
@@ -2,16 +2,8 @@ import nacl from "tweetnacl"
 
 import { WalletClient } from "viem"
 import { getAddress, bytesToHex, hexToBytes, keccak256 } from "viem/utils"
-import { blake3 } from "@noble/hashes/blake3"
-import { equals } from "uint8arrays"
-import { base58btc } from "multiformats/bases/base58"
-import { decode, encode } from "microcbor"
 
-import * as Messages from "./messages.js"
-import { PublicUserRegistration } from "./PublicUserRegistration.js"
-import { getRoomId } from "./RoomRegistration.js"
-import { type KeyBundle, type PrivateUserRegistration, type Room, RoomEvent } from "./types.js"
-import { assert } from "./utils.js"
+import { type KeyBundle, type PrivateUserRegistration } from "./types.js"
 
 const buildMagicString = (pin: string) => `[Password: ${pin}]
 
@@ -101,65 +93,4 @@ export const createPrivateUserRegistration = async (
 		encryptionPrivateKey: privateKey,
 		signingPrivateKey: bytesToHex(derivedSecrets.signingKeyPair.secretKey),
 	}
-}
-
-export const encryptAndSignMessageForRoom = (room: Room, message: string, user: PrivateUserRegistration) => {
-	const event = {
-		type: "message",
-		detail: { content: message, sender: user.address, timestamp: Date.now() },
-	}
-
-	const otherRoomMembers = room.members.filter(({ address }) => user.address !== address)
-	assert(otherRoomMembers.length > 0, "room has no other members")
-
-	const encryptedData = Messages.EncryptedEvent.encode({
-		recipients: otherRoomMembers.map((otherRoomMember) => {
-			const publicKey = hexToBytes(otherRoomMember.keyBundle.encryptionPublicKey)
-			const nonce = nacl.randomBytes(nacl.box.nonceLength)
-			const ciphertext = nacl.box(encode(event), nonce, publicKey, hexToBytes(user.encryptionPrivateKey))
-
-			return {
-				publicKey,
-				ciphertext,
-				nonce,
-			}
-		}),
-		roomId: base58btc.baseDecode(room.id),
-		senderAddress: hexToBytes(user.address),
-		senderPublicKey: hexToBytes(user.keyBundle.encryptionPublicKey),
-	})
-
-	const signature = nacl.sign.detached(encryptedData, hexToBytes(user.signingPrivateKey))
-
-	return Messages.SignedData.encode({ signature, data: encryptedData })
-}
-
-export const decryptEvent = (encryptedEvent: Messages.EncryptedEvent, user: PrivateUserRegistration) => {
-	let messageToDecrypt: Messages.EncryptedPayload
-	let publicKey: Uint8Array
-
-	if (equals(encryptedEvent.senderAddress, hexToBytes(user.address))) {
-		// this user is the sender
-		// decrypt an arbitrary message, so choose the first one
-		messageToDecrypt = encryptedEvent.recipients[0]
-		publicKey = messageToDecrypt.publicKey
-	} else {
-		// otherwise this user is one of the recipients
-		const retrievedMessageToDecrypt = encryptedEvent.recipients.find(({ publicKey }) =>
-			equals(publicKey, hexToBytes(user.keyBundle.encryptionPublicKey))
-		)
-		assert(retrievedMessageToDecrypt !== undefined, "failed to find encrypted message for this user")
-		messageToDecrypt = retrievedMessageToDecrypt
-		publicKey = encryptedEvent.senderPublicKey
-	}
-
-	const decryptedEvent = nacl.box.open(
-		messageToDecrypt.ciphertext,
-		messageToDecrypt.nonce,
-		publicKey,
-		hexToBytes(user.encryptionPrivateKey)
-	)
-	assert(decryptedEvent !== null, "failed to decrypt room event")
-
-	return decode(decryptedEvent) as RoomEvent
 }
